[#34] Move component files to component folder
