+++ src/front/component/Blind.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './style/Blind.scss';
+import '../style/Blind.scss';
 
 function Blind({ children }) {
   return (
