+++ src/front/component/JobView.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import * as DEFS from '../defs';
-import style from './style/JobView.module.scss';
+import * as DEFS from '../../defs';
+import style from '../style/JobView.module.scss';
 const ipcRenderer = window.ipcRenderer;
 
 const { STATE, printSize } = DEFS.default;
